refactor: image_embedding directory
