+++ src/image_embedding/mod.rs
@@ -1,0 +1,9 @@
+use crate::models::image_embedding::ImageEmbeddingModel;
+const DEFAULT_BATCH_SIZE: usize = 256;
+const DEFAULT_EMBEDDING_MODEL: ImageEmbeddingModel = ImageEmbeddingModel::ClipVitB32;
+
+mod init;
+pub use init::*;
+
+mod models;
+pub use models::*;
