+++ src/image_embedding/models.rs
@@ -1,71 +1,28 @@
-use crate::{
-    common::{normalize, Compose, Transform, TransformData, DEFAULT_CACHE_DIR},
-    models::image_embedding::{models_list, ImageEmbeddingModel},
-    Embedding, ModelInfo,
-};
-use anyhow::{anyhow, Result};
-use hf_hub::{
-    api::sync::{ApiBuilder, ApiRepo},
-    Cache,
-};
-use ndarray::{Array3, ArrayView3};
-use ort::{ExecutionProviderDispatch, GraphOptimizationLevel, Session, Value};
-use rayon::{iter::ParallelIterator, slice::ParallelSlice};
 use std::{
     fmt::Display,
     path::{Path, PathBuf},
     thread::available_parallelism,
 };
-const DEFAULT_BATCH_SIZE: usize = 256;
-const DEFAULT_EMBEDDING_MODEL: ImageEmbeddingModel = ImageEmbeddingModel::ClipVitB32;
 
-/// Options for initializing the ImageEmbedding model
-#[derive(Debug, Clone)]
-pub struct ImageInitOptions {
-    pub model_name: ImageEmbeddingModel,
-    pub execution_providers: Vec<ExecutionProviderDispatch>,
-    pub cache_dir: PathBuf,
-    pub show_download_progress: bool,
-}
+use hf_hub::{
+    api::sync::{ApiBuilder, ApiRepo},
+    Cache,
+};
+use ndarray::{Array3, ArrayView3};
+use ort::{GraphOptimizationLevel, Session, Value};
 
-impl Default for ImageInitOptions {
-    fn default() -> Self {
-        Self {
-            model_name: DEFAULT_EMBEDDING_MODEL,
-            execution_providers: Default::default(),
-            cache_dir: Path::new(DEFAULT_CACHE_DIR).to_path_buf(),
-            show_download_progress: true,
-        }
-    }
-}
+use crate::{
+    common::{normalize, Compose, Transform, TransformData},
+    models::image_embedding::models_list,
+    Embedding, ImageEmbeddingModel, ModelInfo,
+};
+use anyhow::anyhow;
+use rayon::prelude::*;
 
-/// Options for initializing UserDefinedImageEmbeddingModel
-///
-/// Model files are held by the UserDefinedImageEmbeddingModel struct
-#[derive(Debug, Clone, Default)]
-pub struct ImageInitOptionsUserDefined {
-    pub execution_providers: Vec<ExecutionProviderDispatch>,
-}
-
-/// Convert ImageInitOptions to ImageInitOptionsUserDefined
-///
-/// This is useful for when the user wants to use the same options for both the default and user-defined models
-impl From<ImageInitOptions> for ImageInitOptionsUserDefined {
-    fn from(options: ImageInitOptions) -> Self {
-        ImageInitOptionsUserDefined {
-            execution_providers: options.execution_providers,
-        }
-    }
-}
-
-/// Struct for "bring your own" embedding models
-///
-/// The onnx_file and preprocessor_files are expecting the files' bytes
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct UserDefinedImageEmbeddingModel {
-    pub onnx_file: Vec<u8>,
-    pub preprocessor_file: Vec<u8>,
-}
+use super::{
+    init::{ImageInitOptions, ImageInitOptionsUserDefined, UserDefinedImageEmbeddingModel},
+    DEFAULT_BATCH_SIZE,
+};
 
 /// Rust representation of the ImageEmbedding model
 pub struct ImageEmbedding {
@@ -89,7 +46,7 @@ impl ImageEmbedding {
     /// Uses the highest level of Graph optimization
     ///
     /// Uses the total number of CPUs available as the number of intra-threads
-    pub fn try_new(options: ImageInitOptions) -> Result<Self> {
+    pub fn try_new(options: ImageInitOptions) -> anyhow::Result<Self> {
         let ImageInitOptions {
             model_name,
             execution_providers,
@@ -130,7 +87,7 @@ impl ImageEmbedding {
     pub fn try_new_from_user_defined(
         model: UserDefinedImageEmbeddingModel,
         options: ImageInitOptionsUserDefined,
-    ) -> Result<Self> {
+    ) -> anyhow::Result<Self> {
         let ImageInitOptionsUserDefined {
             execution_providers,
         } = options;
@@ -161,7 +118,7 @@ impl ImageEmbedding {
         model: ImageEmbeddingModel,
         cache_dir: PathBuf,
         show_download_progress: bool,
-    ) -> Result<ApiRepo> {
+    ) -> anyhow::Result<ApiRepo> {
         let cache = Cache::new(cache_dir);
         let api = ApiBuilder::from_cache(cache)
             .with_progress(show_download_progress)
@@ -191,7 +148,7 @@ impl ImageEmbedding {
         &self,
         images: Vec<S>,
         batch_size: Option<usize>,
-    ) -> Result<Vec<Embedding>> {
+    ) -> anyhow::Result<Vec<Embedding>> {
         // Determine the batch size, default if not specified
         let batch_size = batch_size.unwrap_or(DEFAULT_BATCH_SIZE);
 
@@ -211,7 +168,7 @@ impl ImageEmbedding {
                             _ => Err(anyhow!("Preprocessor configuration error!")),
                         }
                     })
-                    .collect::<Result<Vec<Array3<f32>>>>()?;
+                    .collect::<anyhow::Result<Vec<Array3<f32>>>>()?;
 
                 // Extract the batch size
                 let inputs_view: Vec<ArrayView3<f32>> =
